+++ mayan/apps/converter/tests/test_transformation_views.py
@@ -26,8 +26,8 @@ class TransformationViewTestCase(
 
     def test_transformation_create_post_view_with_permission(self):
         self.grant_access(
-            obj=self._test_document,
-            permission=self._test_layer.permissions['create']
+            obj=self._test_transformation_object_parent,
+            permission=self._test_layer_permission_create
         )
         transformation_count = LayerTransformation.objects.count()
 
@@ -60,8 +60,8 @@ class TransformationViewTestCase(
 
     def test_transformation_create_get_view_with_permission(self):
         self.grant_access(
-            obj=self._test_document,
-            permission=self._test_layer.permissions['create']
+            obj=self._test_transformation_object_parent,
+            permission=self._test_layer_permission_create
         )
         transformation_count = LayerTransformation.objects.count()
 
@@ -96,8 +96,8 @@ class TransformationViewTestCase(
     def test_transformation_delete_view_with_access(self):
         self._create_test_transformation()
         self.grant_access(
-            obj=self._test_document,
-            permission=self._test_layer.permissions['delete']
+            obj=self._test_transformation_object_parent,
+            permission=self._test_layer_permission_delete
         )
         transformation_count = LayerTransformation.objects.count()
 
@@ -133,8 +133,8 @@ class TransformationViewTestCase(
     def test_transformation_edit_view_with_access(self):
         self._create_test_transformation()
         self.grant_access(
-            obj=self._test_document,
-            permission=self._test_layer.permissions['edit']
+            obj=self._test_transformation_object_parent,
+            permission=self._test_layer_permission_edit
         )
         transformation_arguments = self._test_transformation.arguments
 
@@ -158,7 +158,7 @@ class TransformationViewTestCase(
 
         response = self._request_transformation_list_view()
         self.assertNotContains(
-            response=response, text=self._test_document.label,
+            response=response, text=str(self._test_transformation_object),
             status_code=404
         )
         self.assertNotContains(
@@ -173,15 +173,15 @@ class TransformationViewTestCase(
     def test_transformation_list_view_with_access(self):
         self._create_test_transformation()
         self.grant_access(
-            obj=self._test_document,
-            permission=self._test_layer.permissions['view']
+            obj=self._test_transformation_object_parent,
+            permission=self._test_layer_permission_view
         )
 
         self._clear_events()
 
         response = self._request_transformation_list_view()
         self.assertContains(
-            response=response, text=self._test_document.label,
+            response=response, text=str(self._test_transformation_object),
             status_code=200
         )
         self.assertContains(
@@ -210,8 +210,8 @@ class TransformationViewTestCase(
 
     def test_transformation_select_get_view_with_access(self):
         self.grant_access(
-            obj=self._test_document,
-            permission=self._test_layer.permissions['select']
+            obj=self._test_transformation_object_parent,
+            permission=self._test_layer_permission_create
         )
         transformation_count = LayerTransformation.objects.count()
 
@@ -244,8 +244,8 @@ class TransformationViewTestCase(
 
     def test_transformation_select_post_view_with_access(self):
         self.grant_access(
-            obj=self._test_document,
-            permission=self._test_layer.permissions['select']
+            obj=self._test_transformation_object_parent,
+            permission=self._test_layer_permission_create
         )
         transformation_count = LayerTransformation.objects.count()