+++ mayan/apps/converter/urls.py
@@ -54,11 +54,11 @@ urlpatterns_transformations = [
         name='transformation_create', view=TransformationCreateView.as_view()
     ),
     url(
-        regex=r'^layers/(?P<layer_name>[-_\w]+)/transformations/(?P<transformation_id>\d+)/delete/$',
+        regex=r'^object/(?P<app_label>[-\w]+)/(?P<model_name>[-\w]+)/(?P<object_id>\d+)/layers/(?P<layer_name>[-_\w]+)/transformations/(?P<transformation_id>\d+)/delete/$',
         name='transformation_delete', view=TransformationDeleteView.as_view()
     ),
     url(
-        regex=r'^layers/(?P<layer_name>[-_\w]+)/transformations/(?P<transformation_id>\d+)/edit/$',
+        regex=r'^object/(?P<app_label>[-\w]+)/(?P<model_name>[-\w]+)/(?P<object_id>\d+)/layers/(?P<layer_name>[-_\w]+)/transformations/(?P<transformation_id>\d+)/edit/$',
         name='transformation_edit', view=TransformationEditView.as_view()
     )
 ]